Lots of changes after meltdown/redo
I needed to rewire everything, so there are new potentiometer mappings.
A variety of tweaks made to the noise behavior (mostly) which wanted
some improved debugging, so add some of that stuff too and turn it
on/off with a constant.
This totally works and was fun to play with!!!